+++ CHANGELOG.md
@@ -28,6 +28,7 @@ All notable changes to Biology Dictionary will be documented in this file.
 - Fixed truncated empty lines in tooltips
 - Fixed spawn structure support
 - Render placeholder if failed to render entity
+- Fixed entity rotation jump when rendering in detail screen
 
 ## Previous Versions